backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets>=13.0
aiosqlite==0.20.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langchain-aws>=0.1.0
browser-use>=0.1.40
playwright
cryptography==42.0.0
requests>=2.32.3
python-multipart==0.0.9
pydantic>=2.7.4
pydantic-settings==2.2.0
httpx>=0.27.0
python-dotenv>=1.0.1
aiofiles==23.2.1
boto3>=1.28.0
botocore>=1.31.0

# Strands Agents SDK
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
